+++ lib/Alien/Xrepo/Base.pm
@@ -1,7 +1,7 @@
 use v5.40;
 use experimental 'class';
 #
-class Alien::Xrepo::Base v0.9.3 {
+class Alien::Xrepo::Base v0.9.4 {
     use Alien::Xrepo;
     use Path::Tiny;
     use Exporter qw[import];
@@ -167,7 +167,7 @@ class Alien::Xrepo::Base v0.9.3 {
         Alien::Xrepo::Base::Builder->new( alien_class => $alien_class, action => $action, argv => \@ARGV )->execute();
     }
     class    #
-        Alien::Xrepo::Base::Builder v0.9.3 {
+        Alien::Xrepo::Base::Builder v0.9.4 {
         use CPAN::Meta;
         use ExtUtils::Install qw[install];
         use ExtUtils::InstallPaths;
@@ -411,7 +411,7 @@ class Alien::Xrepo::Base v0.9.3 {
         }
         };
     class    #
-        Alien::Xrepo::Base::Alt v0.9.3 {
+        Alien::Xrepo::Base::Alt v0.9.4 {
         field $base : param;
         field $pkg  : param;
         method package_names ()      { $base->package_names }